Removed intMillis() because, rollovers...

+++ DelayedKickOut.h
@@ -18,7 +18,7 @@ public:
 	virtual void checkDelayedActivation();
 	void activateDelayed();
 private:
-	unsigned int activationStart = 0;
+	unsigned long activationStart = 0;
 	unsigned int maxDelay = 2000;
 
 };
