+++ LedPalmController.h
@@ -22,7 +22,7 @@ private:
 	SmartLed* leds[5];
 	uint8_t multipliers[5];
 	uint8_t currentLed = 0;
-	unsigned int nextTime = 0;
+	unsigned long nextTime = 0;
 	unsigned int animationDelay = 15000;
 };
